requirements.txt
python-dotenv>=1.0
requests>=2.31
anthropic>=0.40
